+++ packages/signers/signer/src/from-raw-signer.ts
@@ -7,6 +7,7 @@ import {
   V14,
   V15,
 } from "@polkadot-api/substrate-bindings"
+import { merkleizeMetadata } from "@polkadot-api/merkleize-metadata"
 
 export function getPolkadotSigner(
   publicKey: Uint8Array,
@@ -57,3 +58,32 @@ export function getPolkadotSigner(
     signBytes: getSignBytes(sign),
   }
 }
+
+const METADATA_IDENTIFIER = "CheckMetadataHash"
+const oneU8 = Uint8Array.from([1])
+
+export const withMetadataHash = (
+  networkInfo: { decimals: number; tokenSymbol: string },
+  base: PolkadotSigner,
+): PolkadotSigner => ({
+  ...base,
+  signTx: async (callData, signedExtensions, metadata, ...rest) =>
+    base.signTx(
+      callData,
+      signedExtensions[METADATA_IDENTIFIER]
+        ? {
+            ...signedExtensions,
+            [METADATA_IDENTIFIER]: {
+              identifier: METADATA_IDENTIFIER,
+              value: oneU8,
+              additionalSigned: mergeUint8(
+                oneU8,
+                merkleizeMetadata(metadata, networkInfo).digest(),
+              ),
+            },
+          }
+        : signedExtensions,
+      metadata,
+      ...rest,
+    ),
+})