+++ packages/signers/signer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- `withMetadataHash` enhancer
+
 ## 0.1.16 - 2025-04-24
 
 ### Fixed